Separate out quantum__qis into QSharp.Foundation
