+++ src/QirRuntime/lib/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(QIR)
+add_subdirectory(QSharpFoundation)
 add_subdirectory(QSharpCore)
 add_subdirectory(Simulators)
 add_subdirectory(Tracer)
