+++ src/QirRuntime/lib/QIR/arrays.cpp
@@ -13,7 +13,7 @@
 #include "QirContext.hpp"
 #include "QirTypes.hpp"
 #include "allocationsTracker.hpp"
-#include "quantum__rt.hpp"
+#include "QirRuntime.hpp"
 
 using namespace Microsoft::Quantum;
