+++ src/QirRuntime/lib/QSharpFoundation/qsharp__foundation_internal.hpp
